eth: LogFilter should not have any opinion on the tags provided
This struct is low-level enough that it should just accurately represent the data provided to it, so it should not be removing block tags that match the defaults.
And even more so, it has the wrong default logic for fromBlock, which should default to "latest" if this logic were to be kept.